% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.R
\name{extractAIC.gw}
\alias{extractAIC.gw}
\title{Extract AIC}
\usage{
\method{extractAIC}{gw}(fit, scale, k = 2, ...)
}
\arguments{
\item{fit}{fitted model, usually the result of a fitter like lm.}

\item{scale}{optional numeric specifying the scale parameter of the model, see scale in step. Currently only used in the "lm" method, where scale specifies the estimate of the error variance, and scale = 0 indicates that it is to be estimated by maximum likelihood.}

\item{k}{numeric specifying the 'weight' of the equivalent degrees of freedom (=: edf) part in the AIC formula.}

\item{...}{further arguments (currently unused in base R).}
}
\description{
Computes the (generalized) Akaike An Information Criterion for a fitted parametric model.
}
\examples{
fit<-gw(goals~played,data=goals)
extractAIC(fit)
}

