% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gw.r
\name{gw.fit}
\alias{gw.fit}
\title{Genaralizing Waring Regression}
\usage{
gw.fit(x, y, weights = NULL, k = NULL, kstart = 1, rostart = 2,
  betastart = NULL, offset = NULL, control = list(),
  method = "L-BFGS-B", hessian = TRUE, intercept = TRUE)
}
\arguments{
\item{x}{vector of covariables}

\item{y}{vector of response variables}

\item{weights}{weights of de data is exists}

\item{k}{optional value for the k parameter. If NULL, it is estimated.}

\item{kstart}{starting value for the k parameter.}

\item{rostart}{starting value for the ro parameter.}

\item{betastart}{starting values for the vector of means.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if more than one is specified their sum is used. See model.offset.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{method}{the method to be used in fitting the model. The default method initially uses non-linear minimization (nlm) and Nelder-Mead optimization (optim) to fit a model which is then re-fitted by "L-BFGS-B" (optim). In this way, SE estimates for all the model parameters are provided. "nlm" and "Nelder-Mead" are also possible values, but they do not provide SE estimates for k and ro.}

\item{hessian}{if TRUE, the hessian of f at the minimum is returned.}

\item{intercept}{a logical value ...}

\item{...}{further arguments.}
}
\value{
gw object
}
\description{
Make a Genaralizing Waring Regression from a formula an a dataset extracting
the response variable and de covariates in no method.... and call to
gw.fit
}
\examples{
gw(goals~played+position,data=goals)
}

