% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gw.r
\name{gw}
\alias{gw}
\title{Genaralizing Waring Regression}
\usage{
gw(formula, data, weights, k = NULL, subset, na.action, kstart = 1,
  rostart = 2, betastart = NULL, offset, control = list(...),
  method = NULL, hessian = TRUE, model = TRUE, x = FALSE, y = TRUE,
  ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula).}

\item{weights}{an optional vector of 'prior weights' to be used in the fitting process. Should be NULL or a numeric vector.}

\item{k}{optional value for the k parameter. If NULL, it is estimated.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. See glm.}

\item{kstart}{starting value for the k parameter.}

\item{rostart}{starting value for the ro parameter.}

\item{betastart}{starting values for the vector of means.}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if more than one is specified their sum is used. See model.offset.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{method}{the method to be used in fitting the model. The default method initially uses non-linear minimization (nlm) and Nelder-Mead optimization (optim) to fit a model which is then re-fitted by "L-BFGS-B" (optim). In this way, SE estimates for all the model parameters are provided. "nlm" and "Nelder-Mead" are also possible values, but they do not provide SE estimates for k and ro.}

\item{hessian}{if TRUE, the hessian of f at the minimum is returned.}

\item{model}{a logical value indicating whether model frame should be included as a component of the returned value.}

\item{x,y}{logical values indicating whether the response vector and model matrix used in the fitting process should be returned as components of the returned value.}

\item{...}{further arguments.}
}
\value{
gw object
}
\description{
Make a Genaralizing Waring Regression from a formula an a dataset extracting
the response variable and de covariates in no method.... and call to
}
\examples{
gw(goals~played+position,data=goals)
}

