\name{GWRM-package}
\alias{GWRM-package}
\alias{GWRM}
\docType{package}
\title{
\packageTitle{GWRM}
}
\description{
\packageDescription{GWRM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GWRM}
\packageIndices{GWRM}
GWRM is a package for fitting and computing Generalized Waring Regression Models. It includes functions for fitting the model to data, for diagnosis and for computing important statistics, as those related to the partition of the variance. It also includes the dataset and the example of Rodriguez-Avi \emph{et al.} (2009).
}
\author{
Maintainer: \packageMaintainer{GWRM}
}

\references{
Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., Olmo-Jimenez, M. J. and Martinez Rodriguez, A. M.(2009). A generalized Waring regression model for count data. Computational Statistics and Data Analysis, 53, pp. 3717-3725
}

\keyword{ package }
\keyword{ count data regression }