% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogLik.R
\name{logLik.gw}
\alias{logLik.gw}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{gw}(object, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "gw".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns an object of class logLik. This is a number with at least one attribute, "df" (degrees of freedom), giving the number of (estimated) parameters in the model.
There is a simple print method for "logLik" objects.
}
\description{
This function is generic; method functions can be written to handle specific classes of objects. Classes which have methods for this function include: "gw"
}
\examples{
object <- gw(goals ~ played, data = goals)
logLik(object)
}

