% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partvar.R
\name{partvar.gw}
\alias{partvar.gw}
\title{partvar gw object.}
\usage{
\method{partvar}{gw}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object class "gw" for which the partition is desired.}

\item{newdata}{optionally, a data frame in which to look for variables with which to obtain the partition. If omitted, all the cases are used.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
partvar gw object
}

