% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partvar.R
\name{partvar}
\alias{partvar}
\title{Partition of the variance in a GWRM model}
\usage{
partvar(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object class "gw" for which the partition is desired.}

\item{newdata}{optionally, a data frame in which to look for variables with which to obtain the partition. If omitted, all the cases are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Two data frames, with ratio of sources of variation and sources of variation in which variance is splitted.
}
\description{
In a GWRM model, the variance may be split into three terms. The first component of this decomposition represents the variability due to randomness and it comes from the underlying Poisson model. The other two components refer to the variability that is not due to randomness but is explained by the presence of liability and proneness, respectively.
}
\examples{
data(goals)
fit <- gw(goals ~ position, data = goals)
pos <- factor(c("Defender", "Midfielder"), levels = c("Defender", "Midfielder", "Forward"))
lev <- data.frame(position = pos, played = c(17, 21))

partvar(fit, newdata = lev)
}

