% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gw.r
\name{print.gw}
\alias{print.gw}
\title{Print a GWRM model}
\usage{
\method{print}{gw}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object class "gw" for which the partition is desired.}

\item{digits}{number ob digit used}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Sintesis de un objeto gw
}
\description{
Print a GWRM model
}
\examples{
data(goals)
fit <- gw(goals ~ position, data = goals)
print(fit)
}

