% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gw.r
\name{print.summary.gw}
\alias{print.summary.gw}
\title{Print Summarizing GWRM model Fits}
\usage{
\method{print}{summary.gw}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object class "summary.gw" for which the partition is desired.}

\item{digits}{number ob digit used}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print summary class for summary class "gw".
}

