% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgw.r
\name{rgw}
\alias{rgw}
\title{Genaralizing Waring}
\usage{
rgw(n, a, k, ro)
}
\arguments{
\item{n}{number of generated values}

\item{a}{a parameter}

\item{k}{k parameter}

\item{ro}{ro parameter}
}
\value{
vector of generated values
}
\description{
Random generation for the Gamma distribution with parameters a, k, ro
}
\examples{
rgw(10,2,2,2)
}

