% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gw.r
\name{summary.gw}
\alias{summary.gw}
\title{Summarizing GWRM model Fits}
\usage{
\method{summary}{gw}(object, ...)
}
\arguments{
\item{object}{an object of class "gw", usually, a result of a call to gw}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Sintesis de un objeto gw
}
\description{
summary method for class "gw".
}
\examples{
data(goals)
fit <- gw(goals ~ position, data = goals)
summary(fit)
}

