% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.gw.r
\name{predict.gw}
\alias{predict.gw}
\title{GWRM Predictions}
\usage{
\method{predict}{gw}(object = NULL, newdata = NULL, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{"gw"}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data frame with newdata and their fitted means.
}
\description{
Obtains predictions from a fitted GWRM object.
}
\examples{
data(goals)
fit <- gw(goals ~ position, data = goals)
predict(fit)
}

