% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.gw}
\alias{residuals.gw}
\title{Extract and Visualize Model Residuals}
\usage{
\method{residuals}{gw}(object, type = "pearson", rep = 19,
  envelope = TRUE, title = "Simulated Envelope of Residuals",
  trace = FALSE, parallel = TRUE, ncores = 2, ...)
}
\arguments{
\item{object}{object of class "gw" holding the fitted model}

\item{type}{type of residuals to be extracted. Default is pearson. response and deviance are also available.}

\item{rep}{number of replications for envelope construction. Default is 19, that is the smallest 95 percent band that can be build}

\item{envelope}{a logical value to specify if the envelope is required.}

\item{title}{a title for the envelope.}

\item{trace}{if TRUE a sort of information is printed during the running time.}

\item{parallel}{if TRUE use parallel executation}

\item{ncores}{is the number of cores that we use if parallel is TRUE}

\item{...}{further arguments passed to or from other methods.}
}
\value{
residuals values and plot
}
\description{
residuals is a method which extracts model residuals from "gw", commonly returned
by gw function. Optionally, it produces a normal plot with simulated envelope of the residual.
}
\examples{
data(goals)
fit0 <- gw(goals ~ position, data = goals)
residuals(fit0, type = "pearson", rep = 19, envelope = TRUE, trace = FALSE, ncores=2)
}

