% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgw.r
\name{rgw}
\alias{rgw}
\title{Simulation og Generalized Waring values}
\usage{
rgw(n, a, k, ro)
}
\arguments{
\item{n}{number of random values to return.}

\item{a}{vector of (non-negative) first parameters.}

\item{k}{vector of (non-negative) second parameters.}

\item{ro}{vector of (non-negative) third parameters.
#'}
}
\value{
\code{rgw} is an auxiliar function which generates random samples from a Generalized Waring distribution to be used in the simulated envelope called by \code{residuals}
}
\description{
Random generation of values from a Generalized Waring distribution with parameters \code{a}, \code{k} and \code{ro}.
}

